refactor rss namespace, add tests
It's about time, sorry for the stupidity on my part not doing tests sooner.

Looking at a lot of different kinds of tests and usability around RSS parsing
at the moment.  Will work on the console part once I feel I have enough to get
RSS reading done and be able to get all the parts of an RSS feed that I want to.

Dealing with a performance bug on getting a list of channel/item/ pieces, and
I'm not sure whether it's because I'm doing it wrong or something.