add function to get individual rss item as map
Add a method and test to get an individual item from a collection of
RSS items and transpose it to a map.  I don't really like the
xml/xml1-> call through and through, and it's really ugly if you're
doing anything more than searching for an individual node.  Once you
have the node, there should be something to transpose it to a more sane
map-type.