add fs library
I thought this would be useful to find out if a string was a file or a
url.  haven't discovered the best way to use it, but I'd like
'rss-slurp to just work with any string--be it file, url, or something
that I can parse and insert a handler for.